+++ src/config.rs
@@ -31,7 +31,6 @@ impl Config {
     }
 }
 
-// this is used to specify the supported languages
 // @extendable: add a new variant here to support a new language
 #[derive(Clone, Debug, PartialEq)]
 pub enum Language {
@@ -41,7 +40,6 @@ pub enum Language {
 impl FromStr for Language {
     type Err = &'static str;
 
-    // used to parse the language from the command line arguments
     // @extendable: add a new match arm here to support a new language
     fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
         match s.to_lowercase().as_str() {
@@ -52,7 +50,6 @@ impl FromStr for Language {
 }
 
 impl Language {
-    // get the name of the language from the enum variant
     // @extendable: add a new match arm here to support a new language
     pub fn name(&self) -> &'static str {
         match self {
@@ -60,7 +57,6 @@ impl Language {
         }
     }
 
-    // get the file extension of the language from the enum variant
     // @extendable: add a new match arm here to support a new language
     pub fn ext(&self) -> &'static str {
         match self {
@@ -69,7 +65,6 @@ impl Language {
     }
 }
 
-// used to specify the configuration for each supported language (e.g. test runner, build command, etc.)
 // @extendable: add a new match arm here to support a new language
 pub fn config(language: Language) -> Config {
     match language {
@@ -83,7 +78,6 @@ pub fn config(language: Language) -> Config {
     }
 }
 
-// this is used to specify what constitutes a failed test for each supported language
 // @extendable: add a new match arm here to support a new language
 pub fn is_test_failed(stderr: &str, language: &Language) -> bool {
     match language {
